Fix pre-existing CI failures: drop Python 3.10, pandas dtype in export tests (#109)
save() methods use contextlib.chdir (3.11+), so the package never worked on
3.10 — bump requires-python to >=3.11 and drop 3.10 from the CI matrix. Cast
wide output to float64 before nulling a column in two export tests, which now
raises (was a FutureWarning) under newer pandas in CI.

+++ tests/test_api.py
@@ -2058,7 +2058,7 @@ class TestDimensionTableExport:
     def test_to_wide_flat_drops_all_null_columns(self):
         import numpy as _np
         dt = self._make_dt()
-        wide = dt.wide()
+        wide = dt.wide().astype('float64')
         n_full = len(wide.columns)
         # Null out the first data column entirely
         wide.iloc[:, 0] = _np.nan
@@ -2071,7 +2071,7 @@ class TestDimensionTableExport:
     def test_create_schema_excludes_all_null_columns(self):
         import numpy as _np
         dt = self._make_dt()
-        wide = dt.wide()
+        wide = dt.wide().astype('float64')
         wide.iloc[:, 0] = _np.nan
         dt.wide = lambda: wide
         flat = dt._to_wide_flat()
